Refactor; import resource
