+++ app/src/androidTest/kotlin/com/itsaky/androidide/helper/OnboardingPermissionsInfoHelper.kt
@@ -1,6 +1,5 @@
 package com.itsaky.androidide.helper
 
-import android.view.accessibility.AccessibilityNodeInfo
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiSelector
 import com.kaspersky.kaspresso.testcases.core.testcontext.TestContext
@@ -17,26 +16,7 @@ fun TestContext<Unit>.passPermissionsInfoSlideWithPrivacyDialog() {
         val d = device.uiDevice
         val btn = d.findObject(UiSelector().text(accept))
         if (btn.waitForExists(2_000)) {
-            // Use accessibility click -- button may be in the gesture exclusion zone
-            val root = InstrumentationRegistry.getInstrumentation().uiAutomation
-                .rootInActiveWindow
-            if (root != null) {
-                val nodes = root.findAccessibilityNodeInfosByText(accept)
-                var clicked = false
-                try {
-                    for (node in nodes) {
-                        if (!clicked && node.text?.toString().equals(accept, ignoreCase = true)) {
-                            clicked = node.performAction(AccessibilityNodeInfo.ACTION_CLICK)
-                        }
-                        node.recycle()
-                    }
-                } finally {
-                    root.recycle()
-                }
-                if (!clicked) {
-                    throw AssertionError("Failed to click '$accept' button via accessibility")
-                }
-            }
+            clickFirstAccessibilityNodeByText(accept)
             d.waitForIdle()
         }
     }
@@ -50,26 +30,15 @@ fun TestContext<Unit>.passPermissionsInfoSlideWithPrivacyDialog() {
             throw AssertionError("NEXT button not found on permissions info slide")
         }
 
-        val uiAutomation = InstrumentationRegistry.getInstrumentation().uiAutomation
-        val root = uiAutomation.rootInActiveWindow
-            ?: throw AssertionError("No active window for accessibility")
-        val nodes = root.findAccessibilityNodeInfosByText("NEXT")
-        var clicked = false
-        try {
-            for (node in nodes) {
+        clickFirstAccessibilityNodeByText(
+            searchText = "NEXT",
+            errorLabel = "NEXT",
+            matchBy = { node ->
                 val desc = node.contentDescription?.toString() ?: ""
                 val text = node.text?.toString() ?: ""
-                if (!clicked && (desc.contains("NEXT", ignoreCase = true) || text.contains("NEXT", ignoreCase = true))) {
-                    clicked = node.performAction(AccessibilityNodeInfo.ACTION_CLICK)
-                }
-                node.recycle()
-            }
-        } finally {
-            root.recycle()
-        }
-        if (!clicked) {
-            throw AssertionError("NEXT button found by UIAutomator but accessibility click failed")
-        }
+                desc.contains("NEXT", ignoreCase = true) || text.contains("NEXT", ignoreCase = true)
+            },
+        )
         d.waitForIdle()
     }
 }
